Update the middleware

+++ api/model/user_activity.go
@@ -1,12 +1,12 @@
 package model
 
 type UserActivity struct {
-	ID           uint   `gorm:"primaryKey"`
-	UserID       *uint  `gorm:"index"` // nullable for guests
-	Method       string `gorm:"size:10"`
-	Path         string `gorm:"size:255"`
-	IP           string `gorm:"size:45"`
-	UserAgent    string `gorm:"size:255"`
+	ID           uint    `gorm:"primaryKey"`
+	UserID       *uint64 `gorm:"index"` // nullable for guests
+	Method       string  `gorm:"size:10"`
+	Path         string  `gorm:"size:255"`
+	IP           string  `gorm:"size:45"`
+	UserAgent    string  `gorm:"size:255"`
 	StatusCode   int
 	ResponseTime int64 // milliseconds
 	TimeStamp
